+++ backend/students.php
@@ -119,8 +119,8 @@ $result = $conn->query($sql);
                                               <td><?php echo $row['address'] ?></td>
                                             <td><?php echo $row['phone'] ?></td>
                                             <td>
-                                                <a  href="edit_students.php?id=<?php echo $row['student_id'] ?>"class="btn btn-warning text-white">Edit</a>
-                                                <a href="delete_student.php?id=<?php echo $row['student_id'] ?>"class="btn btn-danger text-white">Delete</a>
+                                                <a href="edit_students.php?id=<?php echo $row['student_id'] ?>"class="btn btn-warning text-white">Edit</a>
+                                                <a onclick="return confirm('are you sure?')" href="delete_student.php?id=<?php echo $row['student_id'] ?>"class="btn btn-danger text-white">Delete</a>
                                                 
                                             </td>
                                         </tr>
